% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_misc.R
\name{gg_aes}
\alias{gg_aes}
\title{Interpret aesthetic mapping into data frame literature}
\usage{
gg_aes(data, mapping = aes())
}
\arguments{
\item{data}{data object that can be used in the ggplot2 package}

\item{mapping}{argument}
}
\description{
This function subset the data specified in the mapping to use aes() in the custom function.
Denote that this is defined for the others.
}
\examples{
gg_aes(ggplot2::mpg, aes(x = displ, y = hwy, colour = fl, group = class))
}
