% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_curve.R
\name{gg_curve}
\alias{gg_curve}
\title{Draw a curve of a function}
\usage{
gg_curve(expr, from = NULL, to = NULL, n = 101, xname = "x",
  xlab = xname, ylab = NULL, xlim = NULL, args = list(), ...)
}
\arguments{
\item{expr}{same argument as [curve()]. The name of a function, or a call or an expression written as a function of x which will evaluate to an object of the same length as x.}

\item{from}{the lower bound of the x-axis}

\item{to}{the upper bound of the x-axis}

\item{n}{the number of x values to be evaluated}

\item{xname}{character; argument of the function in expr, the name of the x axis}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{xlim}{Optionally, restrict the range of the function}

\item{args}{list of the additional argument of the function}

\item{...}{Additional arguments in stat_function()}
}
\description{
Draw a line corresponding to a function expression.
This implements [ggplot2::stat_function()] given a tibble of range [from, to].
}
\examples{
gg_curve(sqrt(1 + x), from = 1, to = 10)
}
