% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\docType{package}
\name{goodgraphic-package}
\alias{goodgraphic-package}
\alias{goodgraphic}
\title{Private tools for ggplot2}
\description{
A set of functions that are frequently used by <\url{https://github.com/ygeunkim}>
}
\author{
Young-geun Kim <\url{https://github.com/ygeunkim}>
}
