% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatter.R
\name{gg_scatter}
\alias{gg_scatter}
\title{Scatter plot with color directed by PCA}
\usage{
gg_scatter(data, mapping, col_low = "#ff4b1f", col_high = "#1fddff")
}
\arguments{
\item{data}{data frame to be plotted}

\item{mapping}{same as ggplot}

\item{col_low}{low value of gradient}

\item{col_high}{high value of gradient}
}
\description{
Draw a scatter plot and color the points along the principal components.
Color direction becomes the data spread the most.
}
\examples{
gg_scatter(mpg, mapping = aes(x = displ, y = hwy))
}
